Rename addons-list to add-ons-list for naming consistency
